Remove custom FlaskGroup cli from app.cli
This disables the `df` command, but I don't think such a thing is
necessary. The reason I put that command there in the first place
was to have a compact way to run the correct app with `<cli> run`
without having to specify the app location each time.

This is easily accomplished with dotenv (.env or .flaskenv), so I'm
removing this more complicated wrinkle.

I updated .gitignore to ignore .flaskenv in addition to .env.
While I was in there, I deleted a bunch of ignored file refs that
are no longer necessary.